similarity_check/requirements.txt
psycopg2-binary
faiss-cpu
sentence-transformers
numpy
google-genai
python-dotenv
